pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "cricket-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.26.0",
    "opencv-python",
    "moviepy",
    "google-generativeai>=0.8.5",
    "openai>=1.97.1",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
]

[project.scripts]
ui = "src.launcher:run_ui"
cricket-ui = "src.launcher:run_ui"
